Move all cert types to sub module and parse more fields
